Move log.Flatten to internal pkg
