+++ log/log.go
@@ -4,6 +4,7 @@ import (
 	"os"
 	"strings"
 
+	"github.com/TykTechnologies/tyk/internal/maps"
 	"github.com/sirupsen/logrus"
 )
 
@@ -24,7 +25,7 @@ func LoadTranslations(thing map[string]interface{}) {
 	formatter.FullTimestamp = true
 	formatter.DisableColors = true
 	log.Formatter = &TranslationFormatter{formatter}
-	translations, _ = Flatten(thing)
+	translations, _ = maps.Flatten(thing)
 }
 
 type TranslationFormatter struct {
